backend/requirements.txt
# FastAPI Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 环境变量管理
python-dotenv>=1.0.0

# 配置加密
cryptography>=41.0.0

# LangChain相关
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# 向量数据库
chromadb>=0.4.0

# PyTorch (根据你的CUDA版本选择)
torch>=2.0.0
transformers>=4.30.0

# HTTP客户端
httpx>=0.25.0
aiohttp>=3.9.0

# PDF处理
pypdf>=3.17.0
pdfplumber>=0.10.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 其他工具
pydantic>=2.0.0
python-docx>=1.0.0
reportlab>=4.0.0
